refactor: improve queue importing
import queue_manager itself insted of importing all queue module
