+++ bot/handlers/admin.py
@@ -8,9 +8,9 @@ from aiogram.types import Message
 from bot import keyboards as kb
 from bot.db import User, get_all_users, update_user_by_id
 from bot.filters import IsAdminFilter
+from bot.utils import queue_manager
 from bot.utils.broadcaster import send, send_queue
 from bot.utils.json_storage import save_bot_settings
-from bot.utils.queue import queue_manager
 from config import settings
 
 router = Router()
