+++ bot/handlers/register.py
@@ -6,10 +6,9 @@ from aiogram.types import Message
 
 from bot import keyboards as kb
 from bot.db import update_user
-from bot.utils.queue import queue_manager
+from bot.utils import queue_manager
 
 router = Router()
-
 
 
 class Register(StatesGroup):
