+++ bot/utils/__init__.py
@@ -1,9 +1,10 @@
-from . import broadcaster, json_storage, queue
+from . import broadcaster, json_storage
 from .create_folder import create_folder
+from .queue import queue_manager
 
 __all__ = [
     "broadcaster",
     "json_storage",
-    "queue",
     "create_folder",
+    "queue_manager",
 ]
